Add option to use destination endpoint as source (#41)
This adds an option to allow the destination endpoint to be used as the source. This seems to be a bug in some devices. It also removes the auto swapping which has problems with some ZWave plus devices that now make use of endpoint 0.

The useDestination=true option enables this.

openhab/openhab2-addons#742 is reverted with this PR.

Closes #16.

Signed-off-by: Chris Jackson chris@cd-jackson.com