put detection of O/D endpoint swap back in.
- Added logging for received encapsulated frames where the destination
endpoint on the controller is > 1, as the controller in the binding has
no endpoints.
- Added a warning for received encapsulated frames where in addition to
the condition above, the originating endpoint is either 0 (root) or 1
(first endpoint on originating node), as this is a firmware bug with
high probability, where originating and destination are swapped.
- Swap only if conditions above are true and option
useDestEndpointAsSource is set. (And don’t log, as this happens at a
lot of frames and since the option is set, it must be intentional).

- Get command class from node root if originating endpoint = 0. Log
this as encapsulating a message that originates from the node root (and
probably is destined for the root or first endpoint of the controller)
is useless.

Signed-off-by: Jan-Willem Spuij jw@spuij.nl (github: jspuij)