“One and DONE” - Zwave battery device optimization
This PR is to optimize the user experience of initializing new Zwave battery devices. The desired result is to be able to put a battery in the shiny new device, follow the device manual inclusion protocol during an Inbox, Zwave “Scan” mode of OpenHab and, in 20 seconds or less, have all the channels displayed and begin to link them to items.  “One and DONE”.
Since my last unmerged battery device PRs I have replaced the fixed 20 second cap in the iterative one second sleep timer with a user defined maximum “hold awake” parameter.  As the user defined default is five seconds, there is no implied maintainer approval of a longer awake duration.  Kind of a “use at your own risk” feature.  Once the device is DONE (or the cap is reached) the sleep message (No More Information-NMI) is created in the device queue.  Setting a higher user defined cap does not mean increased battery usage.  With one second intervals, versus 5 seconds in today’s binding, battery device awake duration is reduced for non-initialized devices (and is the same for initialized).
I also reverted to using the “is Initialization Complete” trigger. With testing, the Device queue can be empty for a few milliseconds between messages while initialization is not complete.
Also added is a kick-Queue action after the one second mark in the Timer Task to get a potentially “stuck” device queue going again. In some scenarios (primarily during initialization, the existing kick-Queue action appears to happen too early (before the event listener sets the device awake). There are examples from the forum in my attached document.  I was able to recreate the “stuck” scenario (by reverting the second “Add Node Stop” & Sleep Timer cap to 5 seconds) and resolve the “stuck” problem with this additional kick-Queue.
Lastly, I have added the shortened timeout for the second calling of the “Add Node Stop” to make this a complete package.  It has the most impact as anything else to “One and DONE”.  The default 5 seconds to cancel this command eats up about 4 seconds of the Timer leaving only one second left in the initial awake interval. Some of the other enhancements in this PR would have less impact, if a full 5 seconds were available for device initialization after device discovery.
I have all the changes in this PR running, some for as long as 8 months (as of Jan 2022). I have 47 nodes with 13 battery nodes of various manufacturers on an Rpi4 with an Rpi3b test system. It is frustrating to see the battery device initialization issues in the forum, as I know they can be resolved better than the standard “reawake many, many times” advice. These changes will save considerable user frustration and many seconds of battery usage.

Signed-off-by: Bob Eckhoff <katmandodo@yahoo.com>