Zwave 700 controller changes
This was extracted from #1765, updated with additional comments, and the removal unneeded file changes. With his PR 700 controllers will work with Openhab. It is not perfect or technically complete, but will buy time and suppress the crazy talk on the forum. I'll add supporting information in the comment section.

Signed-off-by: Bob Eckhoff <katmandodo@yahoo.com>

Bridge Application message Class file?

After several rounds of testing it does not appear the class file is needed.  All of the bridge responses are handled in the transaction manager.  The same appears to be true for the Application Command Message Class, but I just left that.  In all the Debug level logs I have looked at (granted not all inclusive), I did not see any instance where the debug messages in these classes appeared.  The may be other reasons they need to be included of which I am unaware, but I'll put this out there for comment. I'll attach a debug startup with the changes in this commit.

Signed-off-by: Bob Eckhoff <katmandodo@yahoo.com>

Remove Controller from Network Heal

I will probably need “re-education” as to why the controller is included in the first step of the network heal. On the surface it only seems to provide the controller neighbors for its XML and feed the network diagram. It looks to me like the controller is kicked out by the stage advancer after the first step because “Controllers aren’t designed to allow communication with their node” (comment line 393). Even with 500 controllers healing the controller creates a lot of network routed errors and unanswered messages.  (See attached-extracted node 1 to node 1 messages only) However, eventually enough messages are ACK’ed so a new controller XML is created.
However, with my testing of the 700 controller this no longer works at all.  I started a network heal with the 700 controller and all the nodes, except the controller were “healed” in a few minutes.  I could see that none of the controller to controller messages were ACK’ed.  Twenty two hours later, about 30 minutes apart the controller is still trying (see this and this).  Anyway excluding the controller resolves the 700 issue and also shortens the 500 controller network heal by 4 minutes.
Also (IMO) without the controller involved, the 500 controller post-heal routings appear less scrambled from where they started (more direct communication between the device and the controller). My speculation is that since mostly multiple hops are ACk’ed (1 -->device --->device--->1) that there is a bias towards the “healed” path being at least one hop. Over time (without network heals) and with individual node heals, direct paths are restored.
I did notice that the frames a 500 controller and 700 controller send to “Find Nodes” are different. (see pictures).
Without neighbors on the controller it will not appear on the network map. It could be restored manually by adding neighbors using the API explorer.

Signed-off-by: Bob Eckhoff <katmandodo@yahoo.com>

Change class Name for better match

Changed the class name to match the API documentation also following the pattern for Application Command handler 0x04

Signed-off-by: Bob Eckhoff <katmandodo@yahoo.com>

Changes so Transaction advance will work with 700

Although the previous versions all worked there were some loose ends.  This 1) does a more accurate strip of the extra byte in the transaction manager 2) modifies the serial message to pick up the dest node from a Bridge message, and 3) set the Bridge (A8) as the expected reply class, so the transaction advance will work like before (no need to set complete if Bridge message).

Only problem right now is this will only work for 700, so is not mergeable unless a separate branch.  Will work on that.. Need to find a way to get a 500/700 switch in the transaction payload.

Signed-off-by: Bob Eckhoff <katmandodo@yahoo.com>