Set a default for controller_sisnode
There have been a couple of issues in the forum recently where new users leave this at zero (it is an advanced param).  Before the rigorous checking introduced by OH, no entry was fixed by the binding.  Now the rigorous check appears to gum that fix up. Since 99% of all installations will start with controller = Node1 and many of the lifeline groups put node1 in the lifeline automatically the risk of issues here is low (IMO).
Signed-off-by: Bob Eckhoff <katmandodo@yahoo.com>