Address multichannel CC with only EP 0 (#1959)
* Address multichannel CC with only EP 0

I think this will work based on your guidance.  I just made the minimum amount of changes. As I understand it only the GET will be returned and processed to the node advancer, so it will finish?
Signed-off-by: Bob Eckhoff <katmandodo@yahoo.com>

* Modified first Multichannel GET

Modified first Multichannel GET for only when refresh is true. It was still looping. Added some logging.

Signed-off-by: Bob Eckhoff <katmandodo@yahoo.com>

---------

Signed-off-by: Bob Eckhoff <katmandodo@yahoo.com>