Change update memory class for newer SDKs
It appears the location of the controller node ID has changed in the newer SDKs.  Forum issues with SDK 7.21 and 7.22 probably to accommodate the 16 bit LR address range. I have to decide about testing since I'd have to buy a new zstick as I don't want to mess with my working version by trying to upgrade.  Will mark this as draft, since you might not like it anyway.
Signed-off-by: Bob Eckhoff <katmandodo@yahoo.com>