Use the Serial API to reset nodeId to 8 bit
The other PR was not going to work without major refactoring to allow both 8 and 16 bit node IDs.  Since the binding can not support LR nodes anyway (S2 and Smart Start are also required), this PR sends a message to use 8bit. It worked on a zooz 800 with SDK 7.22 originally using 16bit. (I used Simplicity studio to change the zooz 700 controller on 7.21.41.
Signed-off-by: Bob Eckhoff <katmandodo@yahoo.com>