Use the Set Serial API (0x0B) to reset nodeId to 8 bits (#1969)
* Use the Serial API to reset nodeId to 8 bit

The other PR was not going to work without major refactoring to allow both 8 and 16 bit node IDs.  Since the binding can not support LR nodes anyway (S2 and Smart Start are also required), this PR sends a message to use 8bit. It worked on a zooz 800 with SDK 7.22 originally using 16bit. (I used Simplicity studio to change the zooz 700 controller on 7.21.41.
Signed-off-by: Bob Eckhoff <katmandodo@yahoo.com>

* Minor edits from testing

Minor edits of the new SetNodeIdType message class.
Signed-off-by: Bob Eckhoff <katmandodo@yahoo.com>

* Revert "Minor edits from testing"

This reverts commit 2176694376e183f845769e5efd64f685d84b939e.
Signed-off-by: Bob Eckhoff <katmandodo@yahoo.com>

* Minor Edits

Minor edits with removed extraneous classpath file.

Signed-off-by: Bob Eckhoff <katmandodo@yahoo.com>

* Add SetNodeIdTypeTest

Added SetNodeIdTypeTest and minor edits.
Signed-off-by: Bob Eckhoff <katmandodo@yahoo.com>

* Clarify constants in SetNodeIdTypeType

Clarified the constants and moved to the top of the class.

Signed-off-by: Bob Eckhoff <katmandodo@yahoo.com>

* Final edits

"final" edits and caps for constants.

Signed-off-by: Bob Eckhoff <katmandodo@yahoo.com>

---------

Signed-off-by: Bob Eckhoff <katmandodo@yahoo.com>